fix: move queries to queries.ts, use shared ActionResult, shadcn Select, aria-labels, PATCH method
