+++ src/app/(site)/cover-letters/cover-letters-page-client.tsx
@@ -96,7 +96,7 @@ export function CoverLettersPageClient({
                 </h3>
                 <div className='flex gap-1'>
                   <Link href={`/cover-letters/${cl.id}/edit`}>
-                    <Button variant='ghost' size='icon' className='size-7'>
+                    <Button variant='ghost' size='icon' className='size-7' aria-label='Edit cover letter'>
                       <Pencil className='size-3.5' />
                     </Button>
                   </Link>
@@ -106,6 +106,7 @@ export function CoverLettersPageClient({
                     className='size-7 text-destructive'
                     onClick={() => handleDelete(cl.id)}
                     disabled={isPending}
+                    aria-label='Delete cover letter'
                   >
                     <Trash2 className='size-3.5' />
                   </Button>
